requirements.txt
streamlit>=1.33,<2.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
google-genai>=1.27.0
google-cloud-texttospeech>=2.27.0
